git push

+++ ShapeDrawer 5.3/MyLine.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using SplashKitSDK;
 
 namespace ShapeDrawer
@@ -11,22 +7,25 @@ namespace ShapeDrawer
     {
         private float _endX;
         private float _endY;
+
         public MyLine(Color clr, float startX, float startY, float endX, float endY) : base(clr)
         {
             X = startX;
             Y = startY;
             _endX = endX;
             _endY = endY;
-           
-
         }
-        public MyLine() : this(Color.RandomRGB(255), 0, 0, 20, 20) { }
+
+        public MyLine() : this(Color.RandomRGB(255), 0, 0, 20, 20)
+        {
+        }
 
         public float EndX
         {
             get { return _endX; }
             set { _endX = value; }
         }
+
         public float EndY
         {
             get { return _endY; }
@@ -41,36 +40,33 @@ namespace ShapeDrawer
             }
             SplashKit.DrawLine(Color, X, Y, _endX, _endY);
         }
+
         public override void DrawOutline()
         {
             SplashKit.DrawCircle(Color.Black, X, Y, 5);
             SplashKit.DrawCircle(Color.Black, _endX, _endY, 5);
         }
+
         public override bool IsAt(Point2D p)
         {
-            // Calculate the distance from the point to the line
-            double distance = Math.Abs((EndY - Y) * p.X - (EndX - X) * p.Y + EndX * Y - EndY * X)
-                            / Math.Sqrt(Math.Pow(EndY - Y, 2) + Math.Pow(EndX - X, 2));
-
-            // Define a tolerance value for how close the point can be to the line
+            double distance = Math.Abs((EndY - Y) * p.X - (EndX - X) * p.Y + EndX * Y - EndY * X) / Math.Sqrt(Math.Pow(EndY - Y, 2) + Math.Pow(EndX - X, 2));
             double tolerance = 5.0; // Adjust as needed
-
-            // Check if the distance is within the tolerance
             return distance <= tolerance;
         }
+
         public override void SaveTo(StreamWriter writer)
         {
             writer.WriteLine("Line");
-            base.SaveTo(writer); // This will write Color, X, and Y
-            writer.WriteLine(EndX); // Write EndX
-            writer.WriteLine(EndY); // Write EndY
+            base.SaveTo(writer);
+            writer.WriteLine(EndX);
+            writer.WriteLine(EndY);
         }
 
         public override void LoadFrom(StreamReader reader)
         {
-            base.LoadFrom(reader); // This reads Color, X, and Y
-            EndX = reader.ReadSingle(); // Read EndX
-            EndY = reader.ReadSingle(); // Read EndY
+            base.LoadFrom(reader);
+            EndX = reader.ReadInteger();
+            EndY = reader.ReadInteger();
         }
     }
 }
